+++ src/main/java/org/realityforge/bazel/depgen/DepgenValidationException.java
@@ -1,0 +1,15 @@
+package org.realityforge.bazel.depgen;
+
+public final class DepgenValidationException
+  extends DepgenException
+{
+  public DepgenValidationException( final String message )
+  {
+    super( message );
+  }
+
+  public DepgenValidationException( final String message, final Throwable cause )
+  {
+    super( message, cause );
+  }
+}